fix: Add missing American English spellings
All words ending with `-ize`/`-yze` were checked. All of the verbs which do not have a `z` spelling are either archaic or simply wrong.

This regex was used to search for words that end with `ise` or `yse` but don't have the same word ending with `ize` or `yze` in the next line: `"([A-Za-z]+?)(i|y)se",\n(?!"(\1)(\2)ze",)`.